[PHYS 2214] Added energy transport in mech. waves
A quick explanation of potential energy density that I elided from the
notes.

U_k = 1/2 * tau * Dx * (dy/dx)^2
    = 1/2 * tau * dy/dx * dy * Dx/dx
    ~ 1/2 * tau * sin * dy
    ~ 1/2 * tau_y * dy

This resembles a force in the y-direction times the displacement which
is work which is energy. And indeed the displacement in the y direction
is the source of the elastic potential energy. However, the 1/2 seems
suspect. The Dx/dx =~ 1 is not real math (at least not as far as I
know). I didn't enter it in the notes as I don't find it too helpful,
and in fact can be harmful. I enterred it here so it won't be lost
forever.